+++ apps/deploy-web/src/types/feature-flags.ts
@@ -1,0 +1,8 @@
+export type FeatureFlag =
+  | "alerts" // keep new line
+  | "anonymous_free_trial"
+  | "notifications_general_alerts_update"
+  | "ui_deployment_closed_alert"
+  | "billing_usage"
+  | "custodial_auto_topup"
+  | "ui_sdl_log_collector_enabled";